requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
google-generativeai==0.3.2
google-cloud-texttospeech==2.16.2
SpeechRecognition==3.10.1
pydub==0.25.1
PyPDF2==3.0.1